chore: Support launch dates, launch counter
